Fix block number type in MMRRootSigned

+++ primitives/src/frame/bridge/relay_authorities.rs
@@ -79,7 +79,7 @@ pub struct AuthoritiesSetSigned<T: EthereumRelayAuthorities> {
 #[derive(Clone, Debug, Eq, PartialEq, Event, Decode)]
 pub struct MMRRootSigned<T: EthereumRelayAuthorities> {
     /// block number
-    pub block_number: u128,
+    pub block_number: u32,
     /// mmr root
     pub mmr_root: H256,
     /// The redeemed balance
